features/ONG4-50/CrearComponenteDetalleActividades: show details activitity is ok
