+++ src/Components/Activities/ActivitiesServices.js
@@ -3,7 +3,7 @@ import axios from "axios";
 const url = "https://ongapi.alkemy.org/api/activities";
 
 const findById = async id => {
-	return await axios.get(url + "/" + id);
+	return await axios.get(`${url}/${id}`);
 };
 
 const findAll = async () => {
