+++ src/Components/Activities/index.js
@@ -1,3 +1,3 @@
-export { ActivityDetails } from './ActivitiesDetails';
+export { ActivityDetails } from './ActivityDetails';
 export { findAll, findById } from './ActivitiesServices';
 export { ActivitiesList } from './ActivitiesList';
